refactor: Redesign approach to drink customization
Compose Drink with two collections of DrinkComponent. One represents the base version of the drink (e.g. Latte can have Milk and Espresso in baseDrinkComponents). The other represents the requested customizations of the drink (e.g. Latte can have Syrup.Type.TOFFEE_NUT in customizedDrinkComponents). DrinkComponent has a boolean field, "customization", that is used in Drink.addDrinkComponent(DrinkComponent) to determine which collection to insert the new DrinkComponent.

Massive changes to Drink hierarchy. The new DrinkComponent hierarchy sows the foundation for an item-availability feature.