refactor: Clean InputPaneFragment.initButtons() and change enum "Type" to subclasses
DrinkMilksInputPaneFragment, DrinkFlavorsInputPaneFragment, and DrinkCustomizationsInputPaneFragment have functioning buttons that gets parsed and displayed via CustomizedDrinkComponentAdapter.

Subclasses of InputPaneFragment and InputPaneTabbedFragment are initialized using number of rows, number of columns, and a list of String for their buttons' text (contained in Menu class).

TLDR: All buttons in milks, flavors, and customizations input panes are functioning (except for NULL and a few hard-coded [discontinued] menu items).